requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
PyYAML==6.0.1
mediapipe==0.10.8
pydantic==2.5.3